// src/app/interfaces/payments_type.ts
export interface payments_type {
    CODCONDPGTO: string;
    NOME: string;
    AVISTA: string;
    CODCCUSTOAVISTA: string;
    CODCCUSTO: string;
    CODTIPODOC: string;
    CODFORMA: string;
    CODFORMAPRAZO: string;
    BUSCADESCONTOCLIENTE: string;
    ACEITACONSUMIDOR: string;
    EXPORTAPALM: string;
    INATIVO: string;
}