// src/app/interfaces/client.ts
export interface client {
    CODEMPRESA: number;
    CODCFO: string;
    CGCCFO: string;
    TELEFONE: string;
    NOMEFANTASIA: string;
    LIMITECREDITO: number;
    DATAULTALTERACAO: Date;
    DATAULTMOVIMENTO: Date;
    DESC_CONCEITO: string;
    VALOR_VENCIDO: number;
    VALOR_ABERTO: number;
}